check logined or not
