+++ templates/admin.html
@@ -2,7 +2,7 @@
 {% block script %}
 <link rel="stylesheet" type="text/css" href="{% static 'css/admin.css' %}"> 
 <script src="{% static 'js/src/logout.js' %}"></script>
-<script src="{% static 'js/src/valid_login.js' %}"></script>
+<script src="{% static 'js/src/not_logined.js' %}"></script>
 {% endblock %}
 
 {% block nav-left %}
