+++ templates/join.html
@@ -2,6 +2,7 @@
 {% block script %}
 <link rel="stylesheet" type="text/css" href="{% static 'css/join.css' %}">
 <script src="{% static 'js/src/join.js' %}"></script>
+<script src="{% static 'js/src/is_logined.js' %}"></script>
 {% endblock %} 
 {% block nav-left %}
 <li class="nav-item">
