+++ templates/login.html
@@ -20,6 +20,7 @@
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-alpha.6/js/bootstrap.min.js" integrity="sha384-vBWWzlZJ8ea9aCX4pEW3rVHjgjt7zpkNpZk+02D9phzyeVkE+jo0ieGizqPLForn"
         crossorigin="anonymous"></script>
     <script src="{% static 'js/src/login.js' %}"></script>
+    <script src="{% static 'js/src/is_logined.js' %}"></script>
     <link rel="stylesheet" type="text/css" href="{% static 'css/login.css' %}">
 </head>
