+++ templates/order_list.html
@@ -1,0 +1,38 @@
+{% extends 'base1.html' %} {% load static %} {% block script %}
+<!-- <script src="https://code.jquery.com/ui/1.12.1/jquery-ui.js" integrity="sha256-T0Vest3yCU7pafRw9r+settMBX6JkKN06dqBnpQ8d30="
+    crossorigin="anonymous"></script> -->
+<link rel="stylesheet" type="text/css" href="{% static 'css/join.css' %}">
+<script src="{% static 'js/src/modify-password.js' %}"></script>
+<script src="{% static 'js/src/profile-display.js' %}"></script>
+<script src="{% static 'js/src/logout.js' %}"></script>
+<script src="{% static 'js/src/valid_login.js' %}"></script>
+{% endblock %} {% block nav-left %}
+<li class="nav-item">
+    <a class="nav-link" href="#">发布</a>
+</li>
+<li class="nav-item">
+    <a class="nav-link" href="#">接单</a>
+</li>
+<li class="nav-item">
+    <a class="nav-link" href="#">交流</a>
+</li>
+<li class="nav-item">
+    <a class="nav-link" href="/about">关于</a>
+</li>
+{% endblock %} {% block nav-right %}
+<li class="nav-item dropdown">
+    <a class="nav-link dropdown-toggle" href="#" id="dropdown01" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">个人信息</a>
+    <div class="dropdown-menu" aria-labelledby="dropdown01">
+        <a class="dropdown-item" href="/profile">修改个人信息</a>
+        <a class="dropdown-item" id="logout">登出</a>
+    </div>
+</li>
+{% endblock %} {% block search %}
+<form class="form-inline my-2 my-lg-0">
+    <input class="form-control mr-sm-2" type="text" placeholder="Search">
+    <button class="btn btn-outline-success my-2 my-sm-0" type="submit">搜索</button>
+</form>
+{% endblock %}
+{% block content %}
+
+{% endblock %}
