+++ templates/profile.html
@@ -5,7 +5,7 @@
 <script src="{% static 'js/src/modify-password.js' %}"></script>
 <script src="{% static 'js/src/profile-display.js' %}"></script>
 <script src="{% static 'js/src/logout.js' %}"></script>
-<script src="{% static 'js/src/valid_login.js' %}"></script>
+<script src="{% static 'js/src/not_logined.js' %}"></script>
 {% endblock %} {% block nav-left %}
 <li class="nav-item">
     <a class="nav-link" href="#">发布</a>
