+++ website/urls.py
@@ -18,12 +18,12 @@ from django.conf.urls import url
 from website import views
 
 urlpatterns = [
-    url('helloworld', views.HelloWorld.as_view()),
     url('index', views.index.as_view()),
     url('login', views.login.as_view()),
     url('user_admin', views.admin.as_view()),
     url('join', views.join.as_view()),
     url('profile', views.profile.as_view()),
     url('about', views.about.as_view()),
+    url('order_list', views.order_list.as_view()),
     url(r'^$', views.index.as_view()),
 ]
