+++ website/views/__init__.py
@@ -1,8 +1,4 @@
 # -*- coding: utf8 -*-
-from .hello_world import HelloWorld
-from .index import index
-from .admin_ import admin
-from .join import join
-from .login import login
-from .profile import profile
-from .about import about
+from .index import index, about
+from .user_info import profile, login, join, admin
+from .order import order_list
