--- website/views/hello_world.py
@@ -1,7 +1,0 @@
-from django.shortcuts import render
-from django.views.generic import TemplateView
-# Create your views here.
-
-
-class HelloWorld(TemplateView):
-    template_name = 'helloworld.html'
