+++ website/views/index.py
@@ -5,3 +5,6 @@ from django.views.generic import TemplateView
 
 class index(TemplateView):
     template_name = 'index.html'
+
+class about(TemplateView):
+    template_name = 'about.html'
