--- website/views/join.py
@@ -1,7 +1,0 @@
-from django.shortcuts import render
-from django.views.generic import TemplateView
-# Create your views here.
-
-
-class join(TemplateView):
-    template_name = 'join.html'
