+++ website/views/order.py
@@ -3,5 +3,5 @@ from django.views.generic import TemplateView
 # Create your views here.
 
 
-class profile(TemplateView):
-    template_name = 'profile.html'
+class order_list(TemplateView):
+    template_name = 'order_list.html'
